move nuxt project to application dir so that easy to maintain by docker
