+++ application/front/assets/variables.scss
