+++ application/front/components/NuxtLogo.vue
