+++ application/front/components/VuetifyLogo.vue
