+++ application/front/layouts/default.vue
