+++ application/front/pages/inspire.vue
