+++ application/front/static/favicon.ico
