+++ application/front/test/NuxtLogo.spec.js
